+++ tests/system/conftest.py
@@ -4,6 +4,7 @@ import sys
 import time
 import glob
 import shutil
+import datetime
 import subprocess
 from pathlib import Path
 
@@ -53,7 +54,8 @@ def pytest_runtest_logfinish(nodeid, location):
 
 def pytest_runtest_logreport(report):
     if report.when == 'call':
-        banner = '\n************ RESULT %s   %s ' % (report.outcome.upper(), report.nodeid)
+        dt = datetime.timedelta(seconds=int(report.duration))
+        banner = '\n************ RESULT %s   %s  took %s  ' % (report.outcome.upper(), report.nodeid, dt)
         banner += '*' * (140 - len(banner))
         if report.outcome == 'passed':
             banner = '\u001b[32;1m' + banner + '\u001b[0m'